Scrapping product names, prices, attributes and images' links  (#15)
* SCRAPER FUNCTIONALITY ONLY

Scraper is now able to scrap header linksfrom the website and save them
to the links.json file. Also it is able to scrap sublinks, but not with
the sublinks structure included. For instance, if the the structure is:

- Link1
    - Sublink1
        - Sublink1.1
        - Sublink1.2
    - Sublink2
        - Sublink2.1
        - Sublink2.2
    - Sublink3

The scraper will scrap the links like this:

- Link1
    - Sublink1
    - Sublink1.1
    - Sublink1.2
    - Sublink2
    - Sublink2.1
    - Sublink2.2
    - Sublink2
    - Sublink3

This is because the scraper is not able to distinguish between the
sublinks and the sublinks structure. This will be fixed in the next
commit probably.

Signed-off-by: mtarnowski <michal.tarnowski03@gmail.com>

* Menu structure scraping done

Menu scraping done. The results are now stored in menu_structure.json file.
The menu saves its structure.

Signed-off-by: mtarnowski <michal.tarnowski03@gmail.com>

* Scraper code changes so it collects categories and subcategories properly.
Added prototype of products scraping

* Scrapping product names, prices, attributes and images' links

---------

Signed-off-by: mtarnowski <michal.tarnowski03@gmail.com>
Co-authored-by: mtarnowski <michal.tarnowski03@gmail.com>
Co-authored-by: Stefan Furmański <stefoo123@outlook.com>